Added Sink interaction & Water fill zone
Currently a basic implementation since we do not have water areas in Cogito yet, however a Water Fill zone may be useful in future in combination with these.
For now it can be useful for aesthetic reasons such as fillable sinks, which is what is currently implemented with the Kitchen Sink.
The sink slowly fills up when tap is turned on, and drains when turned off.